separate out irc message code
